Remove register address adjustment
I2C registers on the devices we use are 7 bits, so the
addition of 128 set a high bit outside of the address.
The color sensor works regardless of the value of this
bit, and the SDK code happens to set it high. However,
the IMU does not return proper data unless the bit is
low.